move getTeamsByOrgIds to its own file
