+++ packages/server/graphql/public/types/helpers/getTeamsByOrgIds.ts
@@ -1,0 +1,12 @@
+import errorFilter from '../../../errorFilter'
+import {DataLoaderWorker} from '../../../graphql'
+
+export const getTeamsByOrgIds = async (
+  orgIds: string[],
+  dataLoader: DataLoaderWorker,
+  includeArchived: boolean
+) => {
+  const teamsByOrgId = (await dataLoader.get('teamsByOrgIds').loadMany(orgIds)).filter(errorFilter)
+  const teams = teamsByOrgId.flat()
+  return includeArchived ? teams : teams.filter(({isArchived}) => !isArchived)
+}
